EphyWebOverviewModel: Don't choose signal marshallers
"The C signal marshaller should always be NULL, in which case the best
marshaller for the given closure type will be chosen by GLib. This may
be an internal marshaller specific to the closure type, or
g_cclosure_marshal_generic, which implements generic conversion of
arrays of parameters to C callback invocations. GLib used to require the
user to write or generate a type-specific marshaller and pass that, but
that has been deprecated in favour of automatic selection of
marshallers."

https://developer.gnome.org/gobject/unstable/howto-signals.html